requirements.txt
pandas==2.0.3
numpy==1.24.3
PyYAML==6.0.1
